requirements.txt
python-telegram-bot>=20.0
skyfield>=1.46
